Move many of the extensions into headers so changes to the clock will apply and add a test
